requirements.txt
aiosmtpd==1.4.6
atpublic==5.1
attrs==25.3.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnslib==0.9.26
dnspython==2.7.0
filelock==3.18.0
Flask==3.1.0
Flask-SocketIO==5.5.1
fsspec==2025.3.2
h11==0.14.0
huggingface-hub==0.30.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
packaging==24.2
pillow==11.1.0
python-engineio==4.11.2
python-socketio==5.12.1
PythonDNS==0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
setuptools==78.1.0
simple-websocket==1.1.0
sympy==1.13.1
tokenizers==0.21.1
torch==2.6.0+cpu
torchaudio==2.6.0+cpu
torchvision==0.21.0+cpu
tqdm==4.67.1
transformers==4.51.0
typing_extensions==4.13.1
urllib3==2.3.0
Werkzeug==3.1.3
wsproto==1.2.0
xmltodict==0.14.2
